requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.7.2
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
tiktoken==0.5.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3